# Know cfg new more optimized strings.split