closes #11653
```
goos: darwin
goarch: arm64
pkg: github.com/erigontech/erigon/turbo/snapshotsync/freezeblocks
cpu: Apple M3 Max
                        │   old.txt    │               new.txt               │
                        │    sec/op    │   sec/op     vs base                │
FindMergeRange/big        10.846m ± 1%   3.373m ± 4%  -68.90% (p=0.000 n=10)
FindMergeRange/big-16      7.673m ± 0%   2.879m ± 1%  -62.48% (p=0.000 n=10)
FindMergeRange/small      19.516m ± 1%   6.684m ± 1%  -65.75% (p=0.000 n=10)
FindMergeRange/small-16   14.212m ± 0%   5.985m ± 1%  -57.89% (p=0.000 n=10)
geomean                    12.33m        4.439m       -63.98%

                        │    old.txt    │               new.txt                │
                        │     B/op      │     B/op      vs base                │
FindMergeRange/big         7.819Mi ± 0%   1.712Mi ± 0%  -78.11% (p=0.000 n=10)
FindMergeRange/big-16      7.833Mi ± 0%   1.717Mi ± 0%  -78.08% (p=0.000 n=10)
FindMergeRange/small      13.837Mi ± 0%   3.379Mi ± 0%  -75.58% (p=0.000 n=10)
FindMergeRange/small-16   13.855Mi ± 0%   3.391Mi ± 0%  -75.52% (p=0.000 n=10)
geomean                    10.41Mi        2.409Mi       -76.86%

                        │   old.txt    │               new.txt               │
                        │  allocs/op   │  allocs/op   vs base                │
FindMergeRange/big         59.38k ± 0%   22.72k ± 0%  -61.73% (p=0.000 n=10)
FindMergeRange/big-16      59.38k ± 0%   22.72k ± 0%  -61.73% (p=0.000 n=10)
FindMergeRange/small      110.32k ± 0%   48.49k ± 0%  -56.04% (p=0.000 n=10)
FindMergeRange/small-16   110.33k ± 0%   48.50k ± 0%  -56.04% (p=0.000 n=10)
geomean                    80.94k        33.20k       -58.99%
```